# money-mate-server/src/main/java/com/example/money_mate_server/dto/BudgetUpdateRequest.java
package com.example.money_mate_server.dto;

import java.time.LocalDate;

public class BudgetUpdateRequest {
    private int inputAmount;
    private LocalDate startDate;
    private LocalDate endDate;

    public int getInputAmount() {
        return inputAmount;
    }

    public void setInputAmount(int inputAmount) {
        this.inputAmount = inputAmount;
    }

    public LocalDate getStartDate() {
        return startDate;
    }

    public void setStartDate(LocalDate startDate) {
        this.startDate = startDate;
    }

    public LocalDate getEndDate() {
        return endDate;
    }

    public void setEndDate(LocalDate endDate) {
        this.endDate = endDate;
    }
}